style(svg): Rename Won > PresentWon
