+++ packages/pancake-uikit/src/components/Svg/Icons/PresentWon.tsx
