+++ packages/pancake-uikit/src/components/Svg/index.tsx
@@ -56,7 +56,8 @@ export { default as RemoveIcon } from "./Icons/Remove";
 export { default as VerifiedIcon } from "./Icons/Verified";
 export { default as ProgressBunny } from "./Icons/ProgressBunny";
 export { default as WaitIcon } from "./Icons/Wait";
-export { default as Won } from "./Icons/Won";
+export { default as Present } from "./Icons/Present";
+export { default as PresentWon } from "./Icons/PresentWon";
 export { default as TeamPlayerIcon } from "./Icons/TeamPlayer";
 export { default as TestnetIcon } from "./Icons/Testnet";
 export { default as Ticket } from "./Icons/Ticket";
